(Re)move stateful methods from activator
